[COM-5] unified tests & added docs
